src/requirement.txt
pyyaml
MySQL-python
id3reader
#pyinotify mac os不可用 
watchdog
